Separate tokenizer and parser files
