+++ parser.py
@@ -6,46 +6,9 @@ import json
 import os
 import re
 import sys
-from collections import namedtuple
 
-import token_specification
 import codetree
-
-Token = namedtuple("Token", ["type", "value"])
-
-
-def flatten(token_dictionary):
-    """
-    Take a two-dimensional dictionary in form "<category>": {"<token name>" : "<regex pattern>"}.
-    Return a one-dimensional dictionary in form {"<token name>": "<regex pattern>"}.
-    """
-    flat_token_dictionary = {}
-
-    for key, val in token_dictionary.items():
-        if isinstance(val, dict):
-            val = [val]
-        if isinstance(val, list):
-            for subdict in val:
-                deeper = flatten(subdict).items()
-                flat_token_dictionary.update({key2: val2 for key2, val2 in deeper})
-        else:
-            flat_token_dictionary[key] = val
-
-    return flat_token_dictionary
-
-
-def compile_regex(flat_token_specification):
-    return re.compile(
-        "|".join([regex_pattern for regex_pattern in flat_token_specification.values()])
-    )
-
-
-def generate_tokens(regex_pattern, text):
-    scanner = regex_pattern.scanner(text)
-    for m in iter(scanner.match, None):
-        tok = Token(m.lastgroup, m.group())
-        if tok.type != "WS":
-            yield tok
+from tokenizer import Token, Tokenizer
 
 
 class Parser:
@@ -60,10 +23,12 @@ class Parser:
     """
 
     def __init__(self):
-        self.regex = compile_regex(flatten(token_specification.tokens))
+        self.regex = Tokenizer.compile_regex(
+            Tokenizer.flatten(Tokenizer.token_specification)
+        )
 
     def parse(self, text):
-        self.tokens = generate_tokens(self.regex, text)
+        self.tokens = Tokenizer.generate_tokens(self.regex, text)
         self.tok = None  # Last symbol consumed
         self.nexttok = None  # Next symbol tokenized
         self._advance()  # Load first lookahead token
